tasks/108 step 1: C-list neutralizations (ARCH-BOUNDARY §C + the two §A ones)
- vmm-backend: HypercallRegs{rax..rdx} -> HypercallFrame{args: [u64; 4]};
  complete_hypercall's parameter is the neutral return slot (ret). Which guest
  registers carry the frame is now per-arch backend knowledge.
- vmm-backend: Exit::Hlt -> Exit::Idle (HLT and WFI are one concept above the
  trait); ExitReason::Hlt -> Idle; ExitCounts.hlt -> idle. vmm-core's
  above-trait mirror TerminalReason::Hlt -> Idle (encoded byte unchanged).
  telemetry's independent ExitCounts mirror is deliberately untouched (its
  NDJSON schema pins its field names; outside this task's surface).
- environment: HostFault::InjectInterrupt vector widened u8 -> u32 (GIC INTIDs
  exceed 8 bits); wire payload is now u32 LE, BLOB_VERSION 5 -> 6, goldens
  updated deliberately. vmm-core stages/applies with x86 range checks: > 255 is
  rejected at stage time (Unsupported) and at apply time (ContractViolation).
- control-proto: CrashKind::TripleFault -> CrashKind::UnrecoverableFault
  (portable crash taxonomy; wire tag byte 0x01 unchanged).
- vtime: VClockConfig::{tsc_hz, tsc_base} -> {guest_hz, guest_base};
  VClock::tsc() -> guest_ticks() (§C naming-only; arithmetic untouched).
  vm-state's VtimeState mirror follows; wire bytes unchanged. The x86 contract
  document's own tsc-hz key (contract/parse.rs, canonical.rs) is x86 vendor
  data and stays.
- campaign-runner: box-mode ISA defaults (kernel image name, determinism
  cmdline, backend) become per-vendor config data (VendorBootConfig, x86-64
  the sole bundle).

State-hash canonical form: unchanged (renames only; the widened wire field is
not hashed). Full workspace suite green (1690 tests); clippy -D warnings native
+ x86_64-unknown-linux-gnu cross-target green; public-api snapshots updated
(Linux-frozen vmm-backend/vmm-core snapshots hand-updated to match).

+++ consonance/vmm-core/tests/live_host_plane.rs
@@ -242,7 +242,7 @@ fn host_plane_record_replay_closure() {
     let m2 = snapshot_vtime + env_u64("HP_M2_OFF", 1_500_000);
     let deadline = snapshot_vtime + env_u64("HP_DELTA", 5_000_000);
     let gpa = env_u64("HP_GPA", 0x0100_0000);
-    let vector = env_u64("HP_VECTOR", 0x60) as u8;
+    let vector = env_u64("HP_VECTOR", 0x60) as u32;
     let corrupt = wire(HostFault::CorruptMemory {
         gpa,
         mask: BitMask(0xD15EA5ED_C0FFEE01),
